% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_print.R
\name{cor_print}
\alias{cor_print}
\title{cor print}
\usage{
cor_print(x)
}
\arguments{
\item{x}{is the year of interest}
}
\value{
the correlation matrix for all numerical variables in the nba data for a given year
}
\description{
cor print
}
\examples{
cor_print(1950)
}
