% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicefunctions_function.R
\name{get_highest_ft}
\alias{get_highest_ft}
\title{Function 1: Get highest free throw}
\usage{
get_highest_ft(x)
}
\arguments{
\item{x}{is the year of interest}
}
\value{
The max number of free throw points made by any player in a given year
}
\description{
Function 1: Get highest free throw
}
\examples{
get_best_player(1950)
}
