% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oldest_player.R
\name{get_oldest_player}
\alias{get_oldest_player}
\title{Function 2: Get oldest player}
\usage{
get_oldest_player(x)
}
\arguments{
\item{x}{is the year of interest}
}
\value{
the name of the oldest player in a given year
}
\description{
Function 2: Get oldest player
}
\examples{
get_oldest_player(1950)
}
