% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicefunctions_function.R
\name{cor_print}
\alias{cor_print}
\title{Function 3: Get correlation matrix}
\usage{
cor_print(x)
}
\arguments{
\item{x}{is the year of interest}
}
\value{
the correlation matrix for all numerical variables in the nba data for a given year
}
\description{
Function 3: Get correlation matrix
}
\examples{
cor_print(1950)
}
